% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l2lm.R
\name{l2lm}
\alias{l2lm}
\title{Fit the L2E Parameters of a Multivariate Linear Model with a Univariate Response}
\usage{
l2lm(X, y, init_x, partial = F, init_w = 1, print_level = 0)
}
\arguments{
\item{X}{a predictor matrix.}

\item{y}{a response vector.}

\item{init_x}{an initial estimate of \code{a}, \code{b}, and \code{sig}. Defaults respectively to the mean response \code{mean(y)}, a vector zeros of length \code{ncol(X)}, and the sample standard deviation \code{sqrt(var(y))}}

\item{partial}{are the residuals modeled by a full density \eqn{\epsilon ~ \mathcal N(0, \sigma^2)} or a partial density \eqn{\epsilon ~ w \mathcal N(0, \sigma^2)} .}

\item{init_w}{an initial estimate of \code{w}. If \code{partial = FALSE} then w is fixed at 1.}

\item{print_level}{value of the argument \code{print.level} from the \code{\link[stats]{nlm}} function}
}
\value{
The function \code{l2lm} returns a list of L2E parameters
\describe{
  \item{a}{the intercept estimate}
  \item{b}{a vector of estimated coefficients}
  \item{sig}{an estimated standard deviation of the residuals}
  \item{w}{the weight of the partial L2E fit; it can be greater than 1}
  \item{res}{the residuals of the model}
}
}
\description{
Fit the L2E Parameters of a Multivariate Linear Model with a Univariate Response
}
\examples{
## simulate linear model with 20\% contamination
X = c(3*(1:50) + rnorm(50, 0,20), rnorm(10, 0, 20))
y = 1:60
l2lm_out = l2lm(X = X, y = y, partial = F)
plot(X,y)
abline(l2lm_out$a, l2lm_out$b)
}

